fix the structure files
