+++ src/redux/chatSelector.ts
@@ -1,0 +1,1 @@
+export const selectMsgsChat = (state: any) => state.chat;
