+++ src/redux/userSelector.ts
@@ -1,0 +1,1 @@
+export const selectUsersOnline = (state: any) => state.user;
